requirements.txt
numpy==1.26.3
jax==0.4.23
jaxlib==0.4.23
opt-einsum==3.3.0
ml-dtypes==0.3.2
scipy==1.12.0
scikit-learn==1.4.1.post1
threadpoolctl==3.3.0
joblib==1.3.2
cycler==0.12.1
sympy==1.12
mpmath==1.3.0
tqdm==4.66.2
matplotlib==3.8.2
six==1.16.0
pyparsing==3.1.1
packaging==23.2
pillow==10.2.0
fonttools==4.47.2
kiwisolver==1.4.5
contourpy==1.2.0
python-dateutil==2.8.2
pandas-2.2.2
mpi4py==3.1.5